[fix] 주문 가격 타입을 long 으로 변경하여 컴파일 오류 수정

+++ src/main/java/camp/woowak/lab/order/domain/vo/OrderItem.java
@@ -5,21 +5,21 @@ import jakarta.persistence.Embeddable;
 @Embeddable
 public class OrderItem {
 	private Long menuId;
-	private int price;
+	private long price;
 	private int quantity;
-	private int totalPrice;
+	private long totalPrice;
 
 	protected OrderItem() {
 	}
 
-	public OrderItem(Long menuId, int price, int quantity) {
+	public OrderItem(Long menuId, long price, int quantity) {
 		this.menuId = menuId;
 		this.price = price;
 		this.quantity = quantity;
 		this.totalPrice = price * quantity;
 	}
 
-	public int getTotalPrice() {
+	public long getTotalPrice() {
 		return totalPrice;
 	}
 }
